Boiler plate code removed from UserDetails Fragment
