+++ app/src/main/java/com/example/randomuserapplication/presentation/ui/UserDetailsFragment.kt
@@ -1,0 +1,22 @@
+package com.example.randomuserapplication.presentation.ui
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.randomuserapplication.R
+
+class UserDetailsFragment : Fragment() {
+
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_user_details, container, false)
+    }
+
+
+}
